Moving the CurrentGameEndpoint and FeaturedGameEndpoint to a common SpectatorEndpoint.
